requirements.txt
imapclient
python-dotenv